requirements.txt
requests
jinja2
boto3
python-dotenv